Create Using a Pipeline
The solution first splits the data into a training set and a test set using the train_test_split function from sklearn.model_selection. The test_size parameter is set to 0.3, meaning that 30% of the data will be used for the test set, and the rest (70%) will be used for the training set. 

The random_state parameter is set to 0 to ensure that the splits you generate are reproducible. Scikit-learn uses random permutations to generate the splits. The random state that you provide is used as a seed to the random number generator. This ensures that the random numbers are generated in the same order.

Next, the pipeline is fitted on the training data using the fit method. This method fits the resampler SMOTE and the estimator Logistic Regression sequentially on the data, transforming the samples and fitting the transformed samples, respectively.

Then, the predict method is called on the X_test data to make predictions. The predict method applies transforms to the data, and then predicts with the final estimator.

Finally, the results are displayed using the classification_report and confusion_matrix functions from sklearn.metrics. The classification_report function builds a text report showing the main classification metrics, including precision, recall, f1-score, and support. The confusion_matrix function computes the confusion matrix to evaluate the accuracy of a classification. 

The confusion matrix is a summary of prediction results on a classification problem. The number of correct and incorrect predictions are summarized with count values and broken down by each class.